Runtime is int now by default and not cast from long

+++ src/main/java/com/xceptance/xlt/api/engine/TimerData.java
@@ -14,7 +14,7 @@ public abstract class TimerData extends AbstractData
     /**
      * The runtime of the request.
      */
-    private long runTime;
+    private int runTime;
     
     /**
      * Indicates whether or not the request was successful.
@@ -61,7 +61,7 @@ public abstract class TimerData extends AbstractData
      * 
      * @return the run time
      */
-    public long getRunTime()
+    public int getRunTime()
     {
         return runTime;
     }
@@ -93,7 +93,7 @@ public abstract class TimerData extends AbstractData
      */
     public void setRunTime()
     {
-        runTime = GlobalClock.getInstance().getTime() - getTime();
+        runTime = (int) (GlobalClock.getInstance().getTime() - getTime());
     }
 
     /**
@@ -104,7 +104,7 @@ public abstract class TimerData extends AbstractData
      */
     public void setRunTime(final long runTime)
     {
-        this.runTime = runTime;
+        this.runTime = (int) runTime;
     }
 
     /**
@@ -131,7 +131,7 @@ public abstract class TimerData extends AbstractData
         super.parseValues(values);
 
         // read and check the values
-        runTime = ParseNumbers.parseLong(values.get(3));
+        runTime = ParseNumbers.parseInt(values.get(3));
 
         if (runTime < 0)
         {
